\name{peaklist}
\alias{peaklist}
\docType{data}
\title{
HRMS peak list
}
\description{
HPLC-HRMS peak list of a sewage treatment plant effluent sample. 
}
\usage{data(peaklist)}
\format{
  A data frame with 11172 observations on the following 3 variables.
  \describe{
    \item{\code{mass}}{Peak m/z}
    \item{\code{intensity}}{Peak intensity}
    \item{\code{rt}}{Peak retention time [min]}
  }
}
\details{
HPLC-ESI-FTMS Thermo Fisher orbitrap, resolution 100.000@400m/z, positive ionization, profile data.
Generated by Thermo Fisher Formulator peak-picking algorithm; contains satellite peaks for some high intensity peaks.
}
\examples{
data(peaklist)
plot(peaklist[,3],peaklist[,1],pch=19,cex=0.5,xlab="Retention time [min]",ylab="m/z")
}


\keyword{datasets}
