\name{plotadduct}
\alias{plotadduct}

\title{
Plot of frequencies and peak intensities of different adducts
}
\description{
Plots absolute frequencies of different adducts detected by \code{\link[nontarget]{adduct.search}} and boxplots the intensity distributions of associated peaks.
}
\usage{
plotadduct(adduct)
}

\arguments{
  \item{adduct}{
	List of type adduct produced by \code{\link[nontarget]{adduct.search}}.
	}
}
\author{
Martin Loos
}

\seealso{
	\code{\link[nontarget]{adduct.search}}	
}
